# Don't create Supplier when decoding byte