I realized we should avoid this too given just how hot the code path is. Over many forks I can confirm there is a significant (not huge) change in the execution time, so I guess the GC profiler somehow doesn't keep track of lambda overhead.

I considered porting over Guava's `Preconditions` handling of `%s` but figured it's not worth it, the tried and trued if statement pattern doesn't seem like much of a regression to me.

After
```
Benchmark                                                    Mode  Cnt    Score    Error   Units
SpanIdBenchmark.getSpanIdBytes                               avgt   45   22.248 ±  0.098   ns/op
SpanIdBenchmark.getSpanIdBytes:·gc.alloc.rate                avgt   45  685.071 ±  3.113  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.alloc.rate.norm           avgt   45   24.000 ±  0.001    B/op
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Eden_Space       avgt   45  682.211 ± 25.920  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Eden_Space.norm  avgt   45   23.901 ±  0.910    B/op
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Old_Gen          avgt   45    0.002 ±  0.001  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Old_Gen.norm     avgt   45   ≈ 10⁻⁴             B/op
SpanIdBenchmark.getSpanIdBytes:·gc.count                     avgt   45  287.000           counts
SpanIdBenchmark.getSpanIdBytes:·gc.time                      avgt   45  111.000               ms
```

Before
```
Benchmark                                                    Mode  Cnt    Score    Error   Units
SpanIdBenchmark.getSpanIdBytes                               avgt   45   26.344 ±  0.101   ns/op
SpanIdBenchmark.getSpanIdBytes:·gc.alloc.rate                avgt   45  578.526 ±  2.254  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.alloc.rate.norm           avgt   45   24.000 ±  0.001    B/op
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Eden_Space       avgt   45  581.850 ± 25.282  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Eden_Space.norm  avgt   45   24.143 ±  1.074    B/op
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Old_Gen          avgt   45    0.002 ±  0.001  MB/sec
SpanIdBenchmark.getSpanIdBytes:·gc.churn.G1_Old_Gen.norm     avgt   45   ≈ 10⁻⁴             B/op
SpanIdBenchmark.getSpanIdBytes:·gc.count                     avgt   45  230.000           counts
SpanIdBenchmark.getSpanIdBytes:·gc.time                      avgt   45   94.000               ms
```