@anuraaga 2021-06-03
@jkwatson I think a lambda that references a variable, e.g., `first` can't be static since it has that state. Not sure about the GC numbers but good indeed to get some cycles back for each single ID byte