@jkwatson 2021-06-03
It can if the field is a parameter to the static function.