@jkwatson 2021-06-04
> Weird wonder where the value actually comes from - guess it must be able to reference outside stack frames never new that's possible. Well, anyways seem to have pulled a few cycles from somewhere.

I personally don't understand how the callsite bytecode works in this case, so it might be in there.